Add Zeitwerk
Breaking changes: the following have been renamed to allow Zeitwerk autoloading:

| Old                                          | New                                            |
|----------------------------------------------|------------------------------------------------|
| `FreightKit::Contact`                        | `FreightKit::Model::Contact`                   |
| `FreightKit::Error`                          | `FreightKit::Error::Base`                      |
| `FreightKit::DocumentNotFoundError`          | `FreightKit::Error::DocumentNotFound`          |
| `FreightKit::ExpiredCredentialsError`        | `FreightKit::Error::ExpiredCredentials`        |
| `FreightKit::HTTPError`                      | `FreightKit::Error::HTTP`                      |
| `FreightKit::InvalidCredentialsError`        | `FreightKit::Error::InvalidCredentials`        |
| `FreightKit::ResponseError`                  | `FreightKit::Error::Response`                  |
| `FreightKit::ShipmentNotFoundError`          | `FreightKit::Error::ShipmentNotFound`          |
| `FreightKit::UnserviceableAccessorialsError` | `FreightKit::Error::UnserviceableAccessorials` |
| `FreightKit::UnserviceableError`             | `FreightKit::Error::Unserviceable`             |
| `FreightKit::Model`                          | `FreightKit::Model::Base`                      |
| `FreightKit::Credential`                     | `FreightKit::Model::Credential`                |
| `FreightKit::DateTime`                       | `FreightKit::Model::DateTime`                  |
| `FreightKit::DocumentResponse`               | `FreightKit::Model::DocumentResponse`          |
| `FreightKit::Label`                          | `FreightKit::Model::Label`                     |
| `FreightKit::Location`                       | `FreightKit::Model::Location`                  |
| `FreightKit::PickupResponse`                 | `FreightKit::Model::PickupResponse`            |
| `FreightKit::Price`                          | `FreightKit::Model::Price`                     |
| `FreightKit::Rate`                           | `FreightKit::Model::Rate`                      |
| `FreightKit::RateResponse`                   | `FreightKit::Model::RateResponse`              |
| `FreightKit::Response`                       | `FreightKit::Model::Response`                  |
| `FreightKit::Shipment`                       | `FreightKit::Model::Shipment`                  |
| `FreightKit::ShipmentEvent`                  | `FreightKit::Model::ShipmentEvent`             |
| `FreightKit::TrackingResponse`               | `FreightKit::Model::TrackingResponse`          |
