+++ lib/freight_kit/error.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module FreightKit
-  class Error < ActiveUtils::ActiveUtilsError; end
+  module Error; end
 end
