+++ lib/freight_kit/error/base.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module FreightKit
+  module Error
+    class Base < ActiveUtils::ActiveUtilsError; end
+  end
+end
