+++ lib/freight_kit/error/document_not_found.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module FreightKit
+  module Error
+    class DocumentNotFound < Base; end
+  end
+end
