+++ lib/freight_kit/error/expired_credentials.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module FreightKit
+  module Error
+    class ExpiredCredentials < Base; end
+  end
+end
