+++ lib/freight_kit/error/invalid_credentials.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module FreightKit
+  module Error
+    class InvalidCredentials < Base; end
+  end
+end
