+++ lib/freight_kit/error/shipment_not_found.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module FreightKit
+  module Error
+    class ShipmentNotFound < Base; end
+  end
+end
