+++ lib/freight_kit/error/unserviceable.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module FreightKit
+  module Error
+    class Unserviceable < Base; end
+  end
+end
