--- lib/freight_kit/errors.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require 'freight_kit/errors/http_error'
-require 'freight_kit/errors/invalid_credentials_error'
-require 'freight_kit/errors/expired_credentials_error'
-require 'freight_kit/errors/document_not_found_error'
-require 'freight_kit/errors/shipment_not_found_error'
-require 'freight_kit/errors/response_error'
-require 'freight_kit/errors/unserviceable_error'
-require 'freight_kit/errors/unserviceable_accessorials_error'
