--- lib/freight_kit/errors/document_not_found_error.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module FreightKit
-  class DocumentNotFoundError < FreightKit::Error; end
-end
