--- lib/freight_kit/errors/expired_credentials_error.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module FreightKit
-  class ExpiredCredentialsError < FreightKit::InvalidCredentialsError; end
-end
