--- lib/freight_kit/errors/invalid_credentials_error.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module FreightKit
-  class InvalidCredentialsError < FreightKit::Error; end
-end
