--- lib/freight_kit/errors/shipment_not_found_error.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module FreightKit
-  class ShipmentNotFoundError < FreightKit::Error; end
-end
