--- lib/freight_kit/errors/unserviceable_error.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module FreightKit
-  class UnserviceableError < FreightKit::Error; end
-end
