+++ lib/freight_kit/model.rb
@@ -1,17 +1,5 @@
 # frozen_string_literal: true
 
 module FreightKit
-  class Model
-    include ActiveModel::AttributeAssignment
-    include ActiveModel::Validations
-
-    def initialize(attributes = {})
-      assign_attributes(attributes)
-    end
-
-    def attributes
-      instance_values.with_indifferent_access
-    end
-    alias_method :to_hash, :attributes
-  end
+  module Model; end
 end
