--- lib/freight_kit/models.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-require 'freight_kit/models/credential'
-require 'freight_kit/models/label'
-require 'freight_kit/models/price'
-require 'freight_kit/models/rate'
-require 'freight_kit/models/shipment'
-
-require 'freight_kit/models/date_time'
-require 'freight_kit/models/location'
-
-require 'freight_kit/models/response'
-require 'freight_kit/models/document_response'
-require 'freight_kit/models/pickup_response'
-require 'freight_kit/models/rate_response'
-require 'freight_kit/models/shipment_event'
-require 'freight_kit/models/tracking_response'
