+++ spec/lib/freight_kit_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe FreightKit do
+  it { expect { Zeitwerk::Loader.eager_load_all }.not_to(raise_error) }
+end
